Don't update cache on yum search.
By default, `yum search` will update the repositories' information.
Pacman does not do this, and requires you to do an -Sy operation first.
Since we've implemented -Sy for rpm-based systems, we can mimic that
behavior more closely by forcing yum to not update the cache when
searching (which is also faster).